backend/requirements.txt
# File: requirements.txt
django>=5.0.0
django-ninja>=1.1.0
celery>=5.3.0
redis>=5.0.0
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
yfinance>=0.2.0
pydantic>=2.0.0

django-cors-headers>=4.0.0
gunicorn>=21.2.0
uvicorn>=0.25.0

# Database & Auth
psycopg2-binary>=2.9.0
dj-database-url>=2.1.0
python-dotenv>=1.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
requests>=2.31.0
